requirements.txt
allure-pytest==2.13.5
beautifulsoup4==4.12.3
Faker==23.2.1
psycopg-binary==3.1.18
python-dotenv==1.0.0
pytest==7.4.4
pytest-order==1.2.0
pytest-xdist==3.5.0
requests==2.31.0
selenium==4.14.0
SQLAlchemy==2.0.27
setuptools==69.1.0
webdriver-manager==4.0.1
